Migrating code to conform to Intel Manual

+++ kernel/mem/heap.cpp
@@ -13,7 +13,7 @@
 
 uint32_t placement_address;
 
-uint32_t px_kmalloc_internal(uint32_t sz, int align, uint32_t *phys) {
+size_t px_kmalloc_internal(size_t sz, int align, uint32_t *phys) {
     // If the address is not already page-aligned
     if (align == true && (placement_address & 0xFFFFF000)) {
         // Align it.
@@ -23,23 +23,23 @@ uint32_t px_kmalloc_internal(uint32_t sz, int align, uint32_t *phys) {
     if (phys) {
         *phys = placement_address;
     }
-    uint32_t tmp = placement_address;
+    size_t tmp = placement_address;
     placement_address += sz;
     return tmp;
 }
 
-uint32_t px_kmalloc_a(uint32_t sz) {
+size_t px_kmalloc_a(size_t sz) {
     return px_kmalloc_internal(sz, true, nullptr);
 }
 
-uint32_t px_kmalloc_p(uint32_t sz, uint32_t *phys) {
+size_t px_kmalloc_p(size_t sz, size_t *phys) {
     return px_kmalloc_internal(sz, false, phys);
 }
 
-uint32_t px_kmalloc_ap(uint32_t sz, uint32_t *phys) {
+size_t px_kmalloc_ap(size_t sz, size_t *phys) {
     return px_kmalloc_internal(sz, true, phys);
 }
 
-uint32_t px_kmalloc(uint32_t sz) {
+size_t px_kmalloc(size_t sz) {
     return px_kmalloc_internal(sz, false, nullptr);
 }
